+++ diesel/src/pg/mod.rs
@@ -6,12 +6,11 @@
 
 pub mod expression;
 pub mod types;
-pub mod upsert;
 
 mod backend;
 mod connection;
 mod metadata_lookup;
-mod query_builder;
+pub(crate) mod query_builder;
 pub(crate) mod serialize;
 mod transaction;
 mod value;
@@ -23,6 +22,10 @@ pub use self::query_builder::DistinctOnClause;
 pub use self::query_builder::PgQueryBuilder;
 pub use self::transaction::TransactionBuilder;
 pub use self::value::PgValue;
+#[doc(hidden)]
+#[cfg(feature = "with-deprecated")]
+#[deprecated(since = "2.0.0", note = "Use `diesel::upsert` instead")]
+pub use crate::upsert;
 
 /// Data structures for PG types which have no corresponding Rust type
 ///