+++ diesel/src/query_builder/insert_statement/mod.rs
@@ -141,7 +141,6 @@ impl<T, U, Op, Ret> InsertStatement<T, U, Op, Ret> {
         }
     }
 
-    #[cfg(feature = "postgres")]
     pub(crate) fn replace_values<F, V>(self, f: F) -> InsertStatement<T, V, Op, Ret>
     where
         F: FnOnce(U) -> V,