+++ diesel/src/query_builder/upsert/on_conflict_actions.rs
@@ -1,20 +1,18 @@
+use crate::backend::{Backend, SupportsOnConflictClause};
 use crate::expression::{AppearsOnTable, Expression};
-use crate::pg::Pg;
 use crate::query_builder::*;
 use crate::query_source::*;
 use crate::result::QueryResult;
-
-/// Represents `excluded.column` in an `ON CONFLICT DO UPDATE` clause.
-pub fn excluded<T>(excluded: T) -> Excluded<T> {
-    Excluded(excluded)
-}
 
 #[doc(hidden)]
 #[derive(Debug, Clone, Copy)]
 pub struct DoNothing;
 
-impl QueryFragment<Pg> for DoNothing {
-    fn walk_ast(&self, mut out: AstPass<Pg>) -> QueryResult<()> {
+impl<DB> QueryFragment<DB> for DoNothing
+where
+    DB: Backend + SupportsOnConflictClause,
+{
+    fn walk_ast(&self, mut out: AstPass<DB>) -> QueryResult<()> {
         out.push_sql(" DO NOTHING");
         Ok(())
     }
@@ -32,11 +30,12 @@ impl<T> DoUpdate<T> {
     }
 }
 
-impl<T> QueryFragment<Pg> for DoUpdate<T>
+impl<DB, T> QueryFragment<DB> for DoUpdate<T>
 where
-    T: QueryFragment<Pg>,
+    DB: Backend + SupportsOnConflictClause,
+    T: QueryFragment<DB>,
 {
-    fn walk_ast(&self, mut out: AstPass<Pg>) -> QueryResult<()> {
+    fn walk_ast(&self, mut out: AstPass<DB>) -> QueryResult<()> {
         out.unsafe_to_cache_prepared();
         if self.changeset.is_noop()? {
             out.push_sql(" DO NOTHING");
@@ -52,11 +51,18 @@ where
 #[derive(Debug, Clone, Copy)]
 pub struct Excluded<T>(T);
 
-impl<T> QueryFragment<Pg> for Excluded<T>
+impl<T> Excluded<T> {
+    pub(crate) fn new(t: T) -> Self {
+        Excluded(t)
+    }
+}
+
+impl<DB, T> QueryFragment<DB> for Excluded<T>
 where
+    DB: Backend + SupportsOnConflictClause,
     T: Column,
 {
-    fn walk_ast(&self, mut out: AstPass<Pg>) -> QueryResult<()> {
+    fn walk_ast(&self, mut out: AstPass<DB>) -> QueryResult<()> {
         out.push_sql("excluded.");
         out.push_identifier(T::NAME)?;
         Ok(())